[bugfix] Prevent is-rfc268-test to be treated as a test
